Add tutorial step

+++ src/func.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <string.h>
 
 #include "func.h"
@@ -8,14 +9,14 @@ bool starts_with(const char* text, const char* prefix)
         return false;
     }
 
-    unsigned long ln_text = strlen(text);
-    unsigned long ln_pref = strlen(prefix);
+    size_t ln_text = strlen(text);
+    size_t ln_pref = strlen(prefix);
 
     if (ln_pref > ln_text) {
         return false;
     }
 
-    for (unsigned long i = 0; i < ln_pref; i++) {
+    for (size_t i = 0; i < ln_pref; i++) {
         if (text[i] != prefix[i]) {
             return false;
         }
